requirements.txt
streamlit==1.51.0
langchain==0.2.14
langchain-community==0.2.12
langchain-google-genai==1.0.3
google-generativeai==0.5.4
google-genai==1.49.0
faiss-cpu==1.8.0.post1
sentence-transformers==3.1.1
pypdf==6.2.0
PyPDF2==3.0.1
python-dotenv==1.0.1
pandas==2.3.3
numpy==1.26.4
requests==2.32.5
tqdm==4.67.1
torch==2.9.0
transformers==4.57.1
scikit-learn==1.7.2
scipy==1.15.3
protobuf==4.25.8